enkle renames og flytting av filer
