+++ src/frontend/barnetilsyn/steg/2-aktivitet/LesMerHvilkenAktivitet.tsx
