+++ src/frontend/barnetilsyn/steg/4-pass-av-dine-barn/barnepassDokumentUtil.ts
@@ -1,5 +1,6 @@
-import { Barn, Barnepass, PassType, ÅrsakBarnepass } from '../../../typer/barn';
+import { Barn } from '../../../typer/barn';
 import { Dokumentasjonsbehov, VedleggstypePassAvBarn } from '../../../typer/skjema';
+import { Barnepass, PassType, ÅrsakBarnepass } from '../../typer/barnepass';
 
 export const oppdaterDokumentasjonsbehovForBarnMedPass = (
     barnMedPass: Barnepass[],
