+++ src/frontend/barnetilsyn/steg/4-pass-av-dine-barn/typer.ts
@@ -1,4 +1,4 @@
-import { Barnepass } from '../../../typer/barn';
+import { Barnepass } from '../../typer/barnepass';
 
 export interface BarnepassIntern extends Partial<Barnepass> {
     ident: string;
