+++ src/frontend/barnetilsyn/steg/6-oppsummering/ArbeidsrettetAktivitet.tsx
