+++ src/frontend/components/RedirectTilStart.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Navigate } from 'react-router-dom';
+import { Navigate } from 'react-router';
 
 import { useSøknad } from '../context/SøknadContext';
 import { hentStartRoute } from '../utils/routeUtils';
