+++ src/frontend/reiseTilSamling/Forside.tsx
@@ -19,11 +19,11 @@ import { loggBesøk, loggSkjemaStartet } from '../api/analytics';
 import { BekreftelseCheckbox } from '../components/BekreftelseCheckbox';
 import { InfoPunktliste } from '../components/InfoPunktliste';
 import { Container } from '../components/Side';
+import { useReiseTilSamlingSøknad } from './context/ReiseTilSamlingSøknadContext';
 import { LocaleHeading } from '../components/Teksthåndtering/LocaleHeading';
 import { LocaleInlineLenke } from '../components/Teksthåndtering/LocaleInlineLenke';
 import { LocalePunktliste } from '../components/Teksthåndtering/LocalePunktliste';
 import { LocaleTekst } from '../components/Teksthåndtering/LocaleTekst';
-import { useReiseTilSamlingSøknad } from '../context/ReiseTilSamlingSøknadContext';
 import { fellesTekster } from '../tekster/felles';
 import { Skjematype } from '../typer/skjematyper';
 import { hentNesteRoute } from '../utils/routeUtils';
